Be more standard-C compliant.
Lets be honest, Gnu **is** the standard now, but to pretend to be
portable, I will occasionally compile with '-std=c17 -pedantic-errors' to
see where I added code that I didn't realize was Gnu-only.  In this case,
'standard' C variadic macros are broken whereas Gnu's are not.  So I
added DBUG1 as a janky work-around DBUG messages whose strings take no
formatting arguments.

+++ inc/lib/helpers.h
@@ -24,7 +24,7 @@ extern int assert_array_pos(int idx, void **arr,
 extern const char *notdir(const char *path);
 /* Why isn't this in stdlib.h? */
 #define container_of(x, type, member) \
-        ((type *)(((void *)(x)) - offsetof(type, member)))
+        ((type *)((uintptr_t)((void *)(x)) - offsetof(type, member)))
 #ifndef ARRAY_SIZE
 # define ARRAY_SIZE(x) (sizeof(x) / sizeof((x)[0]))
 #endif
